avoid lost wakeup
when use condition variable, an extra predicate variable must
be used to avoid lost wakeup, and the predicate must be guaranteed
by the lock, and even the atomic is used for the predicate, the lock
is still required. considering the sequence:(1),(2),(3). the signal
will issued before thread 2 entering waiting state, and then thread
2 will sleep forever.