145. Binary Tree Postorder Traversal
enque root first, and then for each iteration, enque right and left if
not leaves, deque when it's leaf or both left and right children of the
node have been dequed. And then set the top of the stack as next
iteration.

use another stack to store the parent nodes which haven't been
processed, enque when first enque it's child in the first stack,
deque when the parent node have been dequed.

1. stack top() operation can only do on non-empty stacks.
2. post order: right->left->parent