// 1195. Fizz Buzz Multithreaded using condition
1. conditon variable doesn't contain condition
   must create a "condtion" first, such as if(condition) wait.
   or dead wait could happen when notify happens before wait.

2. condition variable must be paired with locks. the locks is used
to guard the atomicty of the required "condtion"

must ensure the locks get locked before wait.  the wait function first
unlock the locks, then begin sleep. when waken up, it will automatically
lock the locks again, you need to consider to unlock mannually
sometimes.

3. there is no need to use locks as well on notify threads if it is
independent of the wait threads, that is, it doesn't have to wait for
the wait thread's signal.

4. if(condtion)-->notify_one; while(condtion)-->notify_all
5. the condtion_variable is not designed for improving the efficiency,
    but for avoiding dead locks when use pure mutex to do the same job.
   also, it is for qurantee certain execute order among threads.

6. you can't wakeup specific thread by using one condtion_variable, but
you can add a "who" flag, let all waiting threads wakes up and check if
it is itself.

but you can use multiple condition variables to wakes up the
specified thread

7. How many mutex required? == the exclusive relations.
   How many conditon variable required? 1, or the numbers of different kind threads you
   want to wakes up.